+++ plume/src/main/kotlin/io/github/plume/oss/util/DiffGraphUtil.kt
@@ -65,7 +65,7 @@ object DiffGraphUtil {
                 }
                 is io.shiftleft.passes.`DiffGraph$Change$SetNodeProperty` -> {
                     val node = change.node()
-                    driver.updateVertexProperty(node.id(), node.label(), change.key(), change.value())
+                    driver.updateVertexProperty((node as Node).id(), node.label(), change.key(), change.value())
                 }
                 else -> logger.warn("Unsupported DiffGraph operation ${change.javaClass} encountered.")
             }