+++ CHANGELOG.md
@@ -9,11 +9,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Changed
 
-- Upgrade ShiftLeft dependencies to 1.3.236
+- Upgrade ShiftLeft dependencies to 1.3.314
 - Removed `Binding` vertices
 - Can now handle new type arguments API of domain classes
 - `Extractor::project` now takes an optional boolean to disable reaching defs calculation
 - `Extractor::projectReachingDefs` now calculates reaching defs separately
+- Removed `IDriver::getProgramStructure` as it's not used by the core extractor
+- `VertexMapper` now handles new default property system 
 
 ## [0.5.12] - 2021-07-30